+++ app/src/main/java/com/brewthings/app/ui/screens/pill/GraphScreen.kt
@@ -122,7 +122,6 @@ fun GraphScreen(
                         selectedIndex = selectedDataIndex,
                         onSelect = onGraphSelect
                     )
-
                 }
             }
             item {
